refactor: remove Tooltip component
