+++ components/CardTooltip.vue
@@ -1,25 +1,17 @@
 <template>
-  <Tooltip :left-offset="290">
-    <template #tooltip>
-      <div class="card-tooltip">
-        <img v-if="cardName" :src="url" :alt="cardName" />
-      </div>
-    </template>
-    <template #content>
-      <slot />
-    </template>
-  </Tooltip>
+  <span @mousemove="mousemove" @mouseout="mouseout">
+    <div v-if="active" class="card-tooltip" :style="{ left, top }">
+      <img v-if="cardName" :src="url" :alt="cardName" />
+    </div>
+    <slot />
+  </span>
 </template>
 
 <script lang="ts">
 import Vue from "vue";
-import Tooltip from "@/components/Tooltip.vue";
 import getExternalCardData from "@/lib/get-external-card-data";
 
 export default Vue.extend({
-  components: {
-    Tooltip,
-  },
   props: {
     cardName: {
       type: String,
@@ -28,14 +20,37 @@ export default Vue.extend({
   },
   data() {
     return {
+      active: false,
+      left: "0px",
+      top: "0px",
       url: getExternalCardData(this.cardName).images.oracle,
     };
+  },
+  methods: {
+    mousemove(event: MouseEvent): void {
+      const displayOnRightSide = window.innerWidth / 2 - event.clientX > 0;
+
+      this.active = true;
+      this.top = event.clientY - 30 + "px";
+
+      if (displayOnRightSide) {
+        this.left = event.clientX + 50 + "px";
+      } else {
+        this.left = event.clientX - 290 + "px";
+      }
+    },
+    mouseout(): void {
+      this.active = false;
+    },
   },
 });
 </script>
 
 <style scoped>
 .card-tooltip {
+  pointer-events: none;
+  position: fixed;
+  z-index: 9000000;
   overflow: hidden;
   border-radius: 4.75% / 3.5%;
   height: 340px;
