+++ test/components/CardTooltip.spec.ts
@@ -34,14 +34,50 @@ describe("CardTooltip", () => {
     });
   });
 
-  it("adds card image as the tooltip", () => {
+  it("is hidden by default", () => {
     const wrapper = shallowMount(CardTooltip, options);
 
+    expect(wrapper.find(".card-tooltip").exists()).toBe(false);
+  });
+
+  it("reveals and hides tooltip on mousemove and mouseout", async () => {
+    const wrapper = shallowMount(CardTooltip, options);
+
+    await wrapper.find("span").trigger("mousemove");
     expect(wrapper.find(".card-tooltip").exists()).toBe(true);
+
     const img = wrapper.find("img");
     expect(img.attributes("alt")).toBe("Sydri");
     expect(img.attributes("src")).toBe(
       "https://c1.scryfall.com/file/oracle.jpg"
     );
+
+    await wrapper.find("span").trigger("mouseout");
+    expect(wrapper.find(".card-tooltip").exists()).toBe(false);
+  });
+
+  it("sets the tooltip position based on the mousemove event and window width", async () => {
+    const wrapper = shallowMount(CardTooltip, options);
+    const halfWindowSize = Math.floor(window.innerWidth / 2);
+    const mouseOnLeftSide = halfWindowSize - 10;
+    const mouseOnRightSide = halfWindowSize + 10;
+
+    await wrapper.find("span").trigger("mousemove", {
+      clientX: mouseOnLeftSide,
+      clientY: 12,
+    });
+    expect(wrapper.find(".card-tooltip").element.style.left).toBe(
+      `${mouseOnLeftSide + 50}px`
+    );
+    expect(wrapper.find(".card-tooltip").element.style.top).toBe("-18px");
+
+    await wrapper.find("span").trigger("mousemove", {
+      clientX: mouseOnRightSide,
+      clientY: 50,
+    });
+    expect(wrapper.find(".card-tooltip").element.style.left).toBe(
+      `${mouseOnRightSide - 290}px`
+    );
+    expect(wrapper.find(".card-tooltip").element.style.top).toBe("20px");
   });
 });
